Nits and a change to weights computaiton after @alekh reviewed the algorithm for training.
